chore: split changeFields action into dedicated actions
